% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_data_availability.R
\name{aws24HourDataStatus}
\alias{aws24HourDataStatus}
\title{Get AWS Status data.}
\usage{
aws24HourDataStatus(url)
}
\arguments{
\item{url}{the URL of the server. Ex: "http://192.168.1.10:8080"}
}
\value{
a list object
}
\description{
Get AWS last 24 hour data availability.
}
