% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_check_variables.R
\name{awsCheckVariablesIDs}
\alias{awsCheckVariablesIDs}
\title{AWS data availability.}
\usage{
awsCheckVariablesIDs(start_time, end_time, aws_ids, variables, url)
}
\arguments{
\item{start_time}{start time, format "YYYY-mm-dd HH:MM". Ex: "2020-11-08 13:00"}

\item{end_time}{end time, format "YYYY-mm-dd HH:MM". Ex: "2020-11-10 21:55"}

\item{aws_ids}{aws IDs, a vector of aws ID. Ex: c("301", "306", "10050037", "000003")}

\item{variables}{a vector of aws variables. Ex: c("RR", "TT", "PRES")}

\item{url}{the URL of the server. Ex: "http://192.168.1.10:8080"}
}
\value{
a list object
}
\description{
Get AWS data availability for a list of AWS.
}
