% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_data_extraction.R
\name{awsGetDataAggregateIDs}
\alias{awsGetDataAggregateIDs}
\title{Get aggregated data.}
\usage{
awsGetDataAggregateIDs(timestep, start_time, end_time, aws_ids, url)
}
\arguments{
\item{timestep}{time step of the data, "hourly", "daily", "pentad", "dekadal" or "monthly"}

\item{start_time}{start time, format "YYYY-mm-dd HH:00"}

\item{end_time}{end time, format "YYYY-mm-dd HH:00"}

\item{aws_ids}{aws ids, a vector of aws id
Example: c("301", "306", "10050037", "000003")}

\item{url}{the URL of the server. Ex: "http://192.168.1.10:8080"}
}
\value{
a list object
}
\description{
Get aggregated data for multiple AWS.
}
