% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_metadata.R
\name{awsGetMetadata}
\alias{awsGetMetadata}
\title{Get AWS metadata.}
\usage{
awsGetMetadata(url)
}
\arguments{
\item{url}{the URL of the server. Ex: "http://192.168.1.10:8080"}
}
\value{
a list object
}
\description{
Get AWS metadata.
}
